#5 Forgot to convert the parallel cpu implementation to uint16 instead of uint8

+++ src/lib/cpp/cpu_seq/diffusion.cc
@@ -5,7 +5,7 @@
 namespace cpu_seq {
 
     #pragma GCC diagnostic ignored "-Wunused-parameter"
-    void diffusion_in_memory(const uint8_t *__restrict__ voxels, const shape_t &N, const float *__restrict__ kernel, const int64_t kernel_size, const int64_t repititions, uint8_t *__restrict__ output) {
+    void diffusion_in_memory(const uint8_t *__restrict__ voxels, const shape_t &N, const float *__restrict__ kernel, const int64_t kernel_size, const int64_t repititions, uint16_t *__restrict__ output) {
         std::cout << "Not implemented" << std::endl;
     }
